Added target definition for Eclipse Neon
This commit adds a target definition for Eclipse Neon.

The graphical editor will not work since the dependency
org.eclipse.sphinx.graphiti.util.DiagramUtil and the dependencies to the
wizards can not be resolved. The API for OpenDialog has also changed.
However, since these parts are non-essential, with this change EATOP can
be started on Neon and it is possible to open and edit EAST-ADL models
with the EAST-ADL Example Editor.

Signed-off-by: Jan-Philipp Steghöfer <ec@splashstudio.com>